VibiPass/requirements.txt
pywebview>=4.4.1
